end commit

+++ src/main/Main.java
@@ -26,7 +26,7 @@ public class Main {
         int lengthOfHuffmanString = huffmanString.length();
         System.out.println("The binary string has " + lengthOfBinaryString + " characters.");
         System.out.println("The huffman string has " + lengthOfHuffmanString + " characters.");
-        System.out.println("Compression Ratio (Huffman/Binary): " + (lengthOfHuffmanString / lengthOfBinaryString) + "%");
+        System.out.println("Compression Ratio (Huffman/Binary): " + ((float)lengthOfHuffmanString / lengthOfBinaryString) + "%");
         scanner.close();
     }
 }
